requirements.txt
discord.py==2.5.2
aiomysql==0.2.0
pyyaml==6.0.2
dotenv==0.9.9